fix: handle infinity case in gamma and lgamma functions and add test cases for the gamma function
Co-authored-by: wutianyao<wutianyao1@huawei.com>



# message auto-generated for no-merge-commit merge:
!816 merge master into master

fix: handle infinity case in gamma and lgamma functions and add test cases for the gamma function

Created-by: wutianyao
Commit-by: wutianyao
Merged-by: ascend-robot
Description: ## 描述
This commit modifies gamma and lgamma functions to handle the case where the input is infinity. Previously, these functions would return incorrect results for infinity inputs. This change ensures that when the input is positive or negative infinity, the gamma function returns positive or negative infinity respectively, and the lgamma function returns positive infinity, which is consistent with the results computed by torch.This change improves the robustness of the gamma and lgamma functions and ensures correct behavior in edge cases.

## checklist
<!--  [ ] 表示选中 -->
- [x] 是否通过本地IDE对代码进行静态检查
- [x] 是否通过本地IDE对代码进行格式化处理
- [x] 是否进行空指针校验
- [x] 是否进行返回值校验
- [x] 是否正确释放new/malloc申请的内存
- [x] 是否充分考虑接口的异常场景
- [x] 是否正确记录错误日志

See merge request: Ascend/triton-ascend!816