feat: enhance MaskState::parseConstant to handle shaped constants of bool, default enable select analysis and disable MaskState::parseDiv
Co-authored-by: wutianyao<wutianyao1@huawei.com>



# message auto-generated for no-merge-commit merge:
!892 merge master into master

feat: enhance MaskState::parseConstant to handle shaped constants of bool, default enable select analysis and disable MaskState::parseDiv

Created-by: wutianyao
Commit-by: wutianyao
Merged-by: ascend-robot
Description: ## 描述
This commit modifies the MaskState::parseConstant function to handle shaped constants of bool correctly, default enable select analysis and temporarily disable MaskState::parseDiv due to incorrect calculation. 

- Modifies the MaskState::parseConstant function to handle shaped constants of bool
When the constant is a shaped tensor and the dtype is bool, the function now extracts the shape information and stores it in the dims and offsets fields of MaskState. This ensures that the function can correctly process both scalar and shaped constants.

- Default enable select analysis
The selection analysis switch is enabled by default.

- Temporarily disable MaskState::parseDiv due to incorrect calculation
Temporarily disables the MaskState::parseDiv function for arith::DivSIOp operations. The function was found to produce incorrect results, and as a result, it is being disable to prevent potential errors in the compilation process.

- Add testcase for divsiop in select analysis
A new test case has been added to expose the error scenario, ensuring that the issue can be easily identified and fixed in the future.

## checklist
- [x] 是否通过本地IDE对代码进行静态检查
- [x] 是否通过本地IDE对代码进行格式化处理
- [x] 是否进行空指针校验
- [x] 是否进行返回值校验
- [x] 是否正确释放new/malloc申请的内存
- [x] 是否充分考虑接口的异常场景
- [x] 是否正确记录错误日志

See merge request: Ascend/triton-ascend!892